add singleClickable option
It can be useful to single click on entries to add them to the other
side. Add an option to do so. When this option is enabled it doesn't
make sense for doubleClickable to be enabled (since you can't
double-click on something that runs away the moment you single-click on
it) so disable doubleClickable when singleClickable is enabled.

# Conflicts:
# 	README.md